Fix test classes
