+++ core/imds/src/test/java/software/amazon/awssdk/imds/internal/DefaultEc2MetadataClientTest.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.imds;
+package software.amazon.awssdk.imds.internal;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
 import static com.github.tomakehurst.wiremock.client.WireMock.equalTo;
@@ -47,12 +47,15 @@ import software.amazon.awssdk.core.retry.backoff.BackoffStrategy;
 import software.amazon.awssdk.core.retry.backoff.FixedDelayBackoffStrategy;
 import software.amazon.awssdk.http.SdkHttpClient;
 import software.amazon.awssdk.http.urlconnection.UrlConnectionHttpClient;
+import software.amazon.awssdk.imds.Ec2MetadataClient;
+import software.amazon.awssdk.imds.Ec2MetadataRetryPolicy;
+import software.amazon.awssdk.imds.MetadataResponse;
 
 /**
  * Unit Tests to test the Ec2Metadata Client functionality
  */
 @RunWith(MockitoJUnitRunner.class)
-public class Ec2MetadataClientTest {
+public class DefaultEc2MetadataClientTest {
 
     private static final String TOKEN_RESOURCE_PATH = "/latest/api/token";
 
@@ -324,9 +327,9 @@ public class Ec2MetadataClientTest {
         Ec2MetadataClient ec2MetadataRequest = Ec2MetadataClient.builder()
                                                                 .endpoint(URI.create("http://localhost:" + mockMetadataEndpoint.port()))
                                                                 .retryPolicy(Ec2MetadataRetryPolicy.builder()
-                                                                                       .backoffStrategy(noWait)
-                                                                                       .numRetries(numRetries)
-                                                                                       .build())
+                                                                                                   .backoffStrategy(noWait)
+                                                                                                   .numRetries(numRetries)
+                                                                                                   .build())
                                                                 .build();
         assertThatThrownBy(() -> ec2MetadataRequest.get("/latest/meta-data/ami-id"))
             .hasMessageContaining("Exceeded maximum number of retries.")
